Changed JobsController redirections
Changed JobsController#create method to redirect to :dashboard on save

Changed JobsController#delete method to redirect to
:dashboard on deletion

Changed JobsController#checkin method so that check-ins are now
ignored if a Job is in 'paused' status. The rationale goes along the
lines that if tracking a Job is 'paused' then it's by definition
not supposed to be processed. So check-ins shouldn't count. I'm
not wedded to this behavior, and this change may revert in the future.